fix: compile metrics collector on windows
