+++ apps/backend/internal/system/metrics/collector.go
@@ -10,7 +10,6 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
-	"syscall"
 	"time"
 )
 
@@ -181,19 +180,6 @@ func memInfoPercent(path string) (float64, error) {
 		available = values["MemFree"]
 	}
 	return (1 - float64(available)/float64(total)) * 100, nil
-}
-
-func diskPercent(path string) (float64, error) {
-	var stat syscall.Statfs_t
-	if err := syscall.Statfs(path, &stat); err != nil {
-		return 0, err
-	}
-	total := stat.Blocks * uint64(stat.Bsize)
-	free := stat.Bavail * uint64(stat.Bsize)
-	if total == 0 {
-		return 0, errors.New("disk total is zero")
-	}
-	return (1 - float64(free)/float64(total)) * 100, nil
 }
 
 func cpuTemp(sysRoot string) (float64, error) {
